refactored pallet
