+++ Cargo.toml
@@ -22,7 +22,7 @@ parity-scale-codec = { version = "3", default-features = false, features = [
 ] }
 clap = { version = "4.3", features = ["derive"] }
 env_logger = "0.10.0"
-envy = "0.4.2"
+envy = "0.3.1"
 ethereum-types = { version = "0.14.1", default-features = false }
 ethers = "2.0.7"
 eyre = "0.6.8"
@@ -66,13 +66,8 @@ sp-core = { default-features = false, git = "https://github.com/paritytech/subst
 eth-types = { git = "https://github.com/ggxchain/pallet-eth2-light-client.git", branch = "polkadot-v0.9.43", default-features = false, features = [
     "eth2",
 ] }
-eth2-ssz = { package = "webb-eth2-ssz", git = "https://github.com/ggxchain/pallet-eth2-light-client.git", branch = "polkadot-v0.9.43", default-features = false }
-merkle-proof = { package = "webb-merkle-proof", git = "https://github.com/ggxchain/pallet-eth2-light-client.git", branch = "polkadot-v0.9.43", default-features = false }
-bls = { package = "webb-bls", git = "https://github.com/ggxchain/pallet-eth2-light-client.git", branch = "polkadot-v0.9.43", default-features = false }
-tree-hash = { package = "webb-tree-hash", git = "https://github.com/ggxchain/pallet-eth2-light-client.git", branch = "polkadot-v0.9.43", default-features = false }
-tree-hash-derive = { package = "webb-tree-hash-derive", git = "https://github.com/ggxchain/pallet-eth2-light-client.git", branch = "polkadot-v0.9.43", default-features = false }
 consensus = { package = "webb-consensus-types", git = "https://github.com/ggxchain/pallet-eth2-light-client.git", branch = "polkadot-v0.9.43", default-features = false }
-eth2-pallet-init = { package = "webb-eth2-pallet-init", git = "https://github.com/ggxchain/pallet-eth2-light-client.git", branch = "polkadot-v0.9.43" }
+pallet-eth2-light-client = { git = "https://github.com/ggxchain/pallet-eth2-light-client.git", branch = "polkadot-v0.9.43", default-features = false }
 webb-proposals = { git = "https://github.com/webb-tools/webb-rs.git", default-features = false, features = [
     "scale",
     "evm",
