--- pallet/src/data/goerli/beacon_header_5258752.json
@@ -1,7 +1,0 @@
-{
-  "slot": "5258752",
-  "proposer_index": 34176,
-  "parent_root": "0xa83e82602a7fb85d739f39f984102a2839ba78077609be7b90b2f975bc0dc45f",
-  "state_root": "0xbc0ff54babe0da9d35281610d6d54b1d63b0aa068d9edffc0cee89ca54d7fe6a",
-  "body_root": "0x3ada633800d8938a42290892e92b9e6725639b971ef1ebff3c676fcfb8c48634"
-}
