+++ types/Cargo.toml
@@ -6,21 +6,24 @@ authors.workspace = true
 repository.workspace = true
 
 [features]
-default = ["merkle-proof"]
+default = []
 debug = []
 merkle-proof = ["dep:cita_trie", "dep:hasher"]
-std = ["serde/std", "serde-big-array"]
+serde = ["dep:serde", "dep:serde-big-array"]
+std = ["serde", "serde?/std"]
 
 [dependencies]
 alloy-rlp.workspace = true
+parity-scale-codec.workspace = true
+scale-info.workspace = true
 bytes.workspace = true
 keccak-hash.workspace = true
 cita_trie = { workspace = true, optional = true }
 hasher = { workspace = true, optional = true }
 ethbloom.workspace = true
 # Optional deps 
-serde = { workspace = true }
-serde-big-array = { workspace = true, optional = true }
+serde = { workspace = true, optional = true, default-features = false }
+serde-big-array = { workspace = true, optional = true, default-features = false }
 
 [dev-dependencies]
 ethers.workspace = true
@@ -29,3 +32,6 @@ test-strategy.workspace = true
 hex-literal.workspace = true
 serde_json.workspace = true
 hex.workspace = true
+cita_trie.workspace = true
+hasher.workspace = true
+types = { path = ".", features = ["merkle-proof"] }
