+++ types/src/lib.rs
@@ -2,7 +2,6 @@ mod receipt;
 pub use receipt::{
     Log, Receipt, ReceiptMerkleProof, ReceiptMerkleProofNode, TransactionReceipt, TxType,
 };
-use serde::{Deserialize, Serialize};
 
 mod primitives;
 pub use primitives::{H160, H256, H64, U256};
@@ -15,7 +14,8 @@ pub use bloom::Bloom;
 
 pub(crate) mod encode;
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct EventProof {
     /// Block corresponding to a [stored block hash][1] in Webb's `pallet-eth2-light-client`.
     /// The hash of this structure is computed using its [rlp][2] representation. In particular, this is the 12th field of `execution_payload`,
