+++ types/src/receipt/trie/extension.rs
@@ -1,11 +1,11 @@
 use alloy_rlp::Encodable;
-use serde::{Deserialize, Serialize};
 
 use crate::{encode, H256};
 
 use super::nibble::Nibbles;
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct ExtensionNode {
     pub prefix: Vec<u8>,
     pub pointer: H256,
