+++ types/src/receipt/trie/leaf.rs
@@ -1,9 +1,9 @@
 use alloy_rlp::{BufMut, Encodable};
-use serde::{Deserialize, Serialize};
 
 use crate::{encode, receipt::trie::nibble::Nibbles, TransactionReceipt};
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct ReceiptLeaf {
     key: Vec<u8>,
     value: Vec<u8>,
